+++ lib/marshal-parser/formatters/symbols/table.rb
@@ -9,9 +9,22 @@ module MarshalParser
         end
 
         def string
+          width = digits_in(@symbols.size)
+
           @symbols.map.with_index do |symbol, i|
-            "%-4d - :%s" % [i, symbol]
+            "%-#{width}d - :%s" % [i, symbol]
           end.join("\n")
+        end
+
+        private def digits_in(n)
+          i = 0
+
+          while n > 0
+            i += 1
+            n /= 10
+          end
+
+          i
         end
       end
     end