+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - Changed output format of the `ast` command and renamed `content` and
   `dump` nodes to `bytes`.
+- Change a symbols table formatting and adjust the indices column width
+  to the maximum index width
 
 ## Fixed